Find where broadcast bugs are - test showcase
